Organize source files (#97)
- Split `index.js` into smaller files
- Split `test.js` into smaller files and put inside `tests`
- Rewrite some tests creating tests dynamically
- `package.json` Add `test:update-snapshots` script
- [ESLint] Enable `func-style` ('declaration') and `prefer-arrow-callback` rules
