+++ .eslintignore
@@ -1,2 +1,2 @@
-testdata
+tests/data
 vendor
