+++ src/index.js
@@ -1,0 +1,4 @@
+const { explore, exploreBundlesAndWriteHtml } = require('./api');
+
+module.exports = explore;
+module.exports.exploreBundlesAndWriteHtml = exploreBundlesAndWriteHtml;
