+++ src/tree-viz.ejs
@@ -6,9 +6,7 @@
   <title><%= bundles[0].name %> - Source Map Explorer</title>
 
   <style>
-    html, body {
-      height: 100%;
-    }
+    html, body { height: 100% }
 
     body {
       font-family: sans-serif;
@@ -17,9 +15,7 @@
       display: flex;
     }
 
-    select, .webtreemap-caption {
-      font-family: sans-serif;
-    }
+    select, .webtreemap-caption { font-family: sans-serif }
 
     #content {
       display: flex;
