+++ tests/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  env: {
+    mocha: true
+  },
+  rules: {
+    'prefer-arrow-callback': 'off',
+  }
+};
