+++ tests/data/foo.min.inline-map.js
