+++ tests/data/foo.min.no-map.js
